Introduced PolicyEnforcerBase and its single-node implementation.
